feat(whispering): recording detail modal and shared row and modal actions
Replace EditRecordingModal with RecordingDetailModal, a single detail
surface reachable by clicking a row's transcript cell
(RecordingTranscriptCell): play back audio, edit the transcript and
metadata, transcribe, transform, download, copy, or delete.

Extract the per-recording actions into TranscribeRecordingButton and
DownloadRecordingButton so the compact row actions and the detail modal
toolbar share one state machine instead of duplicating the transcribe
liveness logic. Rename row-actions/ to actions/ to reflect that these
controls are now shared by the row and the modal.
