--- apps/whispering/src/routes/(app)/(config)/recordings/row-actions/EditRecordingModal.svelte
@@ -1,266 +1,0 @@
-<script lang="ts">
-	import { InstantString } from '@epicenter/field';
-	import { Button } from '@epicenter/ui/button';
-	import { confirmationDialog } from '@epicenter/ui/confirmation-dialog';
-	import { Input } from '@epicenter/ui/input';
-	import { Label } from '@epicenter/ui/label';
-	import * as Modal from '@epicenter/ui/modal';
-	import { Spinner } from '@epicenter/ui/spinner';
-	import { Textarea } from '@epicenter/ui/textarea';
-	import { TimezoneCombobox } from '@epicenter/ui/timezone-combobox';
-	import EditIcon from '@lucide/svelte/icons/pencil';
-	import { createQuery } from '@tanstack/svelte-query';
-	import { onDestroy } from 'svelte';
-	import { deleteRecordingsWithConfirmation } from '$lib/operations/recordings';
-	import { report } from '$lib/report';
-	import { rpc } from '$lib/rpc';
-	import { services } from '$lib/services';
-	import { type Recording, recordings } from '$lib/state/recordings.svelte';
-
-	let { recording }: { recording: Recording } = $props();
-
-	/**
-	 * Capture the recording ID at setup time for use in cleanup.
-	 *
-	 * Reactive props ($props) can become undefined during Svelte's teardown
-	 * when the parent's data source is deleted (e.g. deleting a recording
-	 * causes the table row, and this component, to unmount). If onDestroy
-	 * reads the prop directly, it may see undefined and throw. Capturing
-	 * the ID here sidesteps the reactive teardown race entirely.
-	 */
-	// svelte-ignore state_referenced_locally -- intentional teardown handle; onDestroy must revoke the original row's URL.
-	const recordingIdForCleanup = recording.id;
-
-	let isDialogOpen = $state(false);
-
-	/**
-	 * A working copy of the recording that we can safely edit.
-	 *
-	 * It's like a photocopy of an important document. You don't want to
-	 * accidentally mess up the original. You edit the photocopy, submit it,
-	 * and the original is updated. Then you get a new photocopy.
-	 *
-	 * Here's how it works:
-	 * 1. We get the original recording data
-	 * 2. We make a copy of it (this variable)
-	 * 3. User makes changes to the copy
-	 * 4. When they save, we send the copy via mutation
-	 * 5. The mutation updates the original recording
-	 * 6. We get the fresh original data back and make a new copy (via $derived)
-	 */
-	let workingCopy = $derived(
-		// Reset the working copy when new recording data comes in.
-		recording,
-	);
-
-	/**
-	 * Tracks whether the user has made changes to the working copy.
-	 *
-	 * Think of this like a "dirty" flag on a document - it tells us if
-	 * the user has made edits that haven't been saved yet.
-	 *
-	 * How it works:
-	 * - Starts as false when we get fresh data from the upstream recording
-	 * - Becomes true as soon as the user edits anything
-	 * - Goes back to false when they save or when fresh data comes in
-	 *
-	 * We use this to:
-	 * - Show confirmation dialogs before closing unsaved work
-	 * - Disable the save button when there's nothing to save
-	 * - Reset the working copy when new data arrives
-	 */
-	let isWorkingCopyDirty = $derived.by(() => {
-		// Reset dirty flag when new recording data comes in
-		recording;
-		return false;
-	});
-
-	/**
-	 * Audio playback URL via TanStack Query.
-	 * Audio blobs are too large for Yjs CRDTs, so they're still served
-	 * from BlobStore. Uses accessor pattern for reactive updates.
-	 */
-	const audioPlaybackUrlQuery = createQuery(
-		() => rpc.audio.getPlaybackUrl(() => recording.id).options,
-	);
-
-	function promptUserConfirmLeave() {
-		if (!isWorkingCopyDirty) {
-			isDialogOpen = false;
-			return;
-		}
-
-		confirmationDialog.open({
-			title: 'Unsaved changes',
-			description: 'You have unsaved changes. Are you sure you want to leave?',
-			confirm: { text: 'Leave' },
-			onConfirm: () => {
-				// Reset working copy and dirty flag
-				workingCopy = recording;
-				isWorkingCopyDirty = false;
-
-				isDialogOpen = false;
-			},
-		});
-	}
-
-	onDestroy(() => {
-		services.blobs.audio.revokeUrl(recordingIdForCleanup);
-	});
-</script>
-
-<Modal.Root bind:open={isDialogOpen}>
-	<Modal.Trigger>
-		{#snippet child({ props })}
-			<Button tooltip="Edit recording" variant="ghost" size="icon" {...props}>
-				<EditIcon class="size-4" />
-			</Button>
-		{/snippet}
-	</Modal.Trigger>
-	<Modal.Content
-		onEscapeKeydown={(e) => {
-			e.preventDefault();
-			if (isDialogOpen) {
-				promptUserConfirmLeave();
-			}
-		}}
-		onInteractOutside={(e) => {
-			e.preventDefault();
-			if (isDialogOpen) {
-				promptUserConfirmLeave();
-			}
-		}}
-	>
-		<Modal.Header>
-			<Modal.Title>Edit recording</Modal.Title>
-			<Modal.Description>
-				Make changes to your recording and click save when you're done.
-			</Modal.Description>
-		</Modal.Header>
-		<div class="space-y-4 p-4">
-			<div class="grid grid-cols-4 items-center gap-4">
-				<Label for="title" class="text-right">Title</Label>
-				<Input
-					id="title"
-					value={workingCopy.title}
-					oninput={(e) => {
-						workingCopy = { ...workingCopy, title: e.currentTarget.value };
-						isWorkingCopyDirty = true;
-					}}
-					class="col-span-3"
-				/>
-			</div>
-			<div class="grid grid-cols-4 items-center gap-4">
-				<Label for="recordedAt" class="text-right">Recorded At</Label>
-				<Input
-					id="recordedAt"
-					value={workingCopy.recordedAt}
-					oninput={(e) => {
-						workingCopy = {
-							...workingCopy,
-							recordedAt: e.currentTarget.value as Recording['recordedAt'],
-						};
-						isWorkingCopyDirty = true;
-					}}
-					class="col-span-3"
-				/>
-			</div>
-			<div class="grid grid-cols-4 items-center gap-4">
-				<Label class="text-right">Recorded Timezone</Label>
-				<div class="col-span-3">
-					<TimezoneCombobox
-						bind:value={() => workingCopy.recordedAtZone,
-							(recordedAtZone) => {
-								workingCopy = {
-									...workingCopy,
-									recordedAtZone: recordedAtZone as Recording['recordedAtZone'],
-								};
-								isWorkingCopyDirty = true;
-							}}
-					/>
-				</div>
-			</div>
-			<div class="grid grid-cols-4 items-center gap-4">
-				<Label for="transcript" class="text-right">Transcript</Label>
-				<Textarea
-					id="transcript"
-					value={workingCopy.transcript}
-					oninput={(e) => {
-						workingCopy = {
-							...workingCopy,
-							transcript: e.currentTarget.value,
-						};
-						isWorkingCopyDirty = true;
-					}}
-					class="col-span-3"
-				/>
-			</div>
-			{#if audioPlaybackUrlQuery.data}
-				<div class="grid grid-cols-4 items-center gap-4">
-					<Label for="audio" class="text-right">Audio</Label>
-					<audio
-						src={audioPlaybackUrlQuery.data}
-						controls
-						class="col-span-3 h-8 w-full"
-					></audio>
-				</div>
-			{/if}
-		</div>
-		<Modal.Footer>
-			<Button
-				onclick={() =>
-					deleteRecordingsWithConfirmation(
-						$state.snapshot(recording),
-						{
-							onSuccess: () => {
-								isDialogOpen = false;
-							},
-						},
-					)}
-				variant="destructive"
-			>
-				Delete
-			</Button>
-			<Button variant="outline" onclick={() => promptUserConfirmLeave()}>
-				Close
-			</Button>
-			<Button
-				onclick={() => {
-					const snapshot = $state.snapshot(workingCopy);
-					if (!InstantString.is(snapshot.recordedAt)) {
-						report.info({
-							title: 'Recorded At is not a valid instant',
-							description:
-								'Use a UTC ISO timestamp like 2026-06-13T16:20:00.000Z.',
-						});
-						return;
-					}
-
-					const { error } = recordings.update(recording.id, {
-						title: snapshot.title,
-						recordedAt: snapshot.recordedAt,
-						recordedAtZone: snapshot.recordedAtZone,
-						transcript: snapshot.transcript,
-					});
-
-					if (error) {
-						report.error({
-							title: 'Could not update recording',
-							cause: error,
-						});
-						return;
-					}
-
-					report.success({
-						title: 'Updated recording!',
-						description: 'Your recording has been updated successfully.',
-					});
-					isDialogOpen = false;
-				}}
-				disabled={!isWorkingCopyDirty}
-			>
-				Save
-			</Button>
-		</Modal.Footer>
-	</Modal.Content>
-</Modal.Root>
